perf: reduce duplicated about dataView
